added static getFragments in the `dolos` class, added `Index` and `Report` interface
