+++ lib/src/bin/tokens.ts
@@ -1,9 +1,9 @@
 #!/usr/bin/env node
-import { CodeTokenizer } from "../lib/tokenizer/codeTokenizer";
+import { CodeTokenizerTreeSitter } from "../lib/tokenizer/codeTokenizerTreeSitter";
 import { File } from "../lib/file/file";
 
 (async () => {
-  const tokenizer = new CodeTokenizer("javascript");
+  const tokenizer = new CodeTokenizerTreeSitter("javascript");
   const stdin = (await File.fromPath("/dev/stdin")).ok();
   for (const { token } of tokenizer.generateTokens(stdin.content)) {
     process.stdout.write(token);
