+++ lib/src/lib/analyze/Index.ts
@@ -1,0 +1,8 @@
+import { Report } from "./report";
+import { File } from "../file/file";
+
+export interface Index {
+    compareFiles(
+        files: File[]
+    ): Promise<Report>;
+}
